requirements.txt
attrs==20.3.0
autobahn==21.3.1
Automat==20.2.0
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
constantly==15.1.0
cryptography==3.4.7
dateparser==1.0.0
google-api-core==1.26.3
google-api-python-client==2.2.0
google-auth==1.29.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
googleapis-common-protos==1.53.0
httplib2==0.19.1
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
influxdb-client==1.16.0
oauthlib==3.1.0
packaging==20.9
protobuf==3.15.8
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyOpenSSL==20.0.1
pyparsing==2.4.7
python-binance==0.7.10
python-dateutil==2.8.1
pytz==2021.1
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
Rx==3.1.1
service-identity==18.1.0
six==1.15.0
Twisted==21.2.0
twisted-iocpsupport==1.0.1
txaio==21.2.1
tzlocal==2.1
ujson==4.0.2
uritemplate==3.0.1
urllib3==1.26.4
zope.interface==5.4.0
